test: split connection path tests out of relay_connectivity.rs
cargo xtask oversized-files rejects relay_connectivity.rs at 1077 lines
(new file over the 1000-line threshold). Move the two connection-path
integration tests to a new tests/connection_path.rs and the shared
wait_for_topic_active_path helper to tests/mod.rs, bringing
relay_connectivity.rs back to 847 lines. No test behavior changes.
